refactor: change location of the files
