+++ src/components/app-header/index.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import AppHeader from "./app-header";
+
+
+export default AppHeader;
