+++ src/components/app/app.css
@@ -1,0 +1,9 @@
+.todo-app {
+    margin: 2rem auto 0 auto;
+    max-width: 400px;
+  }
+  
+  .top-panel {
+    margin: 1rem 0;
+  }
+
