+++ src/components/app/app.tsx
@@ -1,0 +1,34 @@
+import React, { Component, ReactElement, ReactHTMLElement } from "react";
+import ReactDOM from "react-dom";
+import { ItoDoItem } from "../../types/types";
+import AppHeader from "../app-header";
+import ItemStatusFilter from "../item-status-filter";
+import SearchPanel from "../search-panel";
+import ToDoList from "../todo-list";
+import "./app.css";
+
+
+
+
+
+function App():JSX.Element{
+
+    const todoData: ItoDoItem[] = [
+        {label: "Learn Angular", important: false, id:1},
+        {label: "Learn Node", important: false, id:2},
+        {label: "Get Job", important: true, id:3}
+    ]
+
+    return(
+        <div className="todo-app">
+            <AppHeader toDo="1" done="3"/>
+            <div className="top-panel d-flex">
+        <SearchPanel />
+        <ItemStatusFilter />
+      </div>
+            <ToDoList todos={todoData}/>
+        </div>
+    )
+}
+
+export default App;
