+++ src/components/app/index.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import App from "./app";
+
+export default App;
