+++ src/components/search-panel/index.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import SearchPanel from "./search-panel";
+
+export default SearchPanel;
