+++ src/components/todo-list/index.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import ToDoList from "./todo-list";
+
+export default ToDoList;
